# internal/swaggerdocs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Private bank API — syncs Fio Bank transaction data into Postgres, matches member payments to their bank transactions, and detects missed dues. Dev-only documentation (see config.EnableSwaggerDocs): every route below requires a Keycloak bearer token unless noted otherwise.",
        "title": "Bank System API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/account": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Requires the manage-bank-accounts role. Excludes the Fio token itself, only whether one is set.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "List bank accounts",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.bankAccountResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Requires the manage-bank-accounts role. fio_token is encrypted at rest and never echoed back.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Register a bank account",
                "parameters": [
                    {
                        "description": "New bank account",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.createBankAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.bankAccountResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "fio_account_id already registered",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/account/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Requires the manage-bank-accounts role. The row and its sync history are kept, just marked inactive.",
                "tags": [
                    "accounts"
                ],
                "summary": "Soft-delete a bank account",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Bank account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "no content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Requires the manage-bank-accounts role. fio_account_id/iban/currency are Fio-assigned and not editable here.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Update a bank account's label and/or Fio token",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Bank account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.updateBankAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.bankAccountResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/account/{id}/backfill": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Requires the manage-bank-accounts role. Pulls via Fio's /periods/ endpoint for a date range predating the account's cursor-based sync; ranges over 90 days old need SCA unlocked in Fio's own Internet Banking first. Runs transaction processing synchronously afterward.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Backfill historical transactions for one account",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Bank account ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Date range, both YYYY-MM-DD, inclusive",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.backfillRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.syncResultResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "sync disabled, or account inactive/has no token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "502": {
                        "description": "Fio API call failed",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.backfillRequest": {
            "type": "object",
            "properties": {
                "from": {
                    "description": "YYYY-MM-DD, inclusive",
                    "type": "string"
                },
                "to": {
                    "description": "YYYY-MM-DD, inclusive",
                    "type": "string"
                }
            }
        },
        "handler.bankAccountResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "display_name": {
                    "type": "string"
                },
                "fio_account_id": {
                    "type": "string"
                },
                "has_token": {
                    "type": "boolean"
                },
                "iban": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                }
            }
        },
...
        "handler.createBankAccountRequest": {
            "type": "object",
            "properties": {
                "currency": {
                    "type": "string"
                },
                "display_name": {
                    "type": "string"
                },
                "fio_account_id": {
                    "type": "string"
                },
                "fio_token": {
                    "type": "string"
                },
                "iban": {
                    "type": "string"
                }
            }
        },
...
        "handler.syncResultResponse": {
            "type": "object",
            "properties": {
                "bank_account_id": {
                    "type": "integer"
                },
                "transactions_failed": {
                    "type": "integer"
                },
                "transactions_fetched": {
                    "type": "integer"
                },
                "transactions_inserted": {
                    "type": "integer"
                },
                "transactions_processed": {
                    "type": "integer"
                }
            }
        },
...
        "handler.updateBankAccountRequest": {
            "type": "object",
            "properties": {
                "display_name": {
                    "type": "string"
                },
                "fio_token": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Keycloak-issued bearer token for bank-system's Keycloak client, sent as `Authorization: Bearer \u003ctoken\u003e`. Which role(s) a given route requires is noted on that route below — see internal/keycloak.Role for the full list.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}